Move the process-group lifecycle where two tools can share it
The reverting check ends the subtree it spawned by group id, and knows that a pid still answering is not the same as a process still running. The pre-commit gate needs both facts for the same reason and currently has neither, so the logic moves to scripts/lib/process-group.js rather than being copied into a second tool. It was subtle enough to be got wrong once already; two versions of it would be two places to get it wrong again.

Verbatim except for one addition: endGroup now takes a grace period. The reverting check keeps the short default, which its own comment justifies by noting that nothing it spawns has a restore step to skip. That is not true of the gate, whose group is a mutation harness holding a real source file rewritten on disk, so it needs to ask for longer.

No behaviour change here. The reverting check's 52 tests, including the eight that drive process groups directly, pass unchanged.
